Refactor `websocket_listener` class.
- adds a context object for storing state that needs to be referenced from within the
callables generated during `__call__()`.
- values that were being calculated within `__call__()`, including values needing to
be resolved through the handler hierarchy, are now referenced inside the generated
callable via the context object.
- the values on the context object are populated in the `on_registration()` callback.
- refactor of the websocket_handlers module into package.
