+++ litestar/handlers/__init__.py
@@ -4,19 +4,19 @@ from .http_handlers import HTTPRouteHandler, delete, get, head, patch, post, put
 from .websocket_handlers import WebsocketListener, WebsocketRouteHandler, websocket, websocket_listener
 
 __all__ = (
-    "asgi",
     "ASGIRouteHandler",
     "BaseRouteHandler",
+    "HTTPRouteHandler",
+    "WebsocketListener",
+    "WebsocketRouteHandler",
+    "asgi",
     "delete",
     "get",
     "head",
-    "HTTPRouteHandler",
     "patch",
     "post",
     "put",
     "route",
     "websocket",
     "websocket_listener",
-    "WebsocketListener",
-    "WebsocketRouteHandler",
 )
