+++ tests/handlers/websocket/test_listeners.py
@@ -239,6 +239,8 @@ def test_listener_callback_no_data_arg_raises() -> None:
         def handler() -> None:
             ...
 
+        handler.on_registration()
+
 
 def test_listener_callback_request_and_body_arg_raises() -> None:
     with pytest.raises(ImproperlyConfiguredException):
@@ -247,8 +249,12 @@ def test_listener_callback_request_and_body_arg_raises() -> None:
         def handler_request(data: str, request: Request) -> None:
             ...
 
+        handler_request.on_registration()
+
     with pytest.raises(ImproperlyConfiguredException):
 
         @websocket_listener("/")
         def handler_body(data: str, body: bytes) -> None:
             ...
+
+        handler_body.on_registration()
